services/detection/requirements.txt
torch==2.1.0
torchvision==0.16.0
opencv-python-headless==4.8.1.78
pillow==10.1.0
redis==5.0.1
boto3==1.34.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
huggingface-hub==0.19.4
ultralytics==8.0.228
